+++ nengo_gui/page.py
@@ -86,6 +86,9 @@ class Page(object):
         self.config_save_time = None   # time of last config file save
         self.config_save_period = 2.0  # minimum time between saves
 
+        self.keys_pressed = set()
+        self.key_codes_pressed = set()
+
         self.lock = threading.Lock()
 
         # use the default filename if none is given
@@ -211,6 +214,7 @@ class Page(object):
         code_locals = {}
         code_locals['nengo_gui'] = nengo_gui
         code_locals['__file__'] = self.filename
+        code_locals['__page__'] = self
 
         self.code = code
         self.error = None